added static to functions

+++ include/kernel/heap.h
@@ -18,7 +18,7 @@ typedef struct
 
 } footer_t;
 
-extern uint32_t kmalloc_page(uint32_t size, int align, uint32_t *physical);
+extern uint32_t kmalloc_page(uint32_t size, uint8_t align, uint32_t *physical);
 extern uint32_t kmalloc(uint32_t size);
 extern uint32_t kmalloc_aligned(uint32_t size);
 extern uint32_t kmalloc_physical(uint32_t size, uint32_t *physical);
