+++ src/kernel/heap.c
@@ -3,7 +3,7 @@
 
 uint32_t heap_address;
 
-uint32_t kmalloc_page(uint32_t size, int align, uint32_t *physical)
+uint32_t kmalloc_page(uint32_t size, uint8_t align, uint32_t *physical)
 {
 
     if (align == 1 && (heap_address & 0xFFFFF000))
